Update.
	* include/netdb.h: Change type of first parameter of __getnetbyaddr_r.
	* inet/getnetbyad.c: Likewise.
	* inet/getnetbyad_r.c: Likewise.
	* nis/nss_nis/nis-network.c: Likewise.
	* nis/nssplus_nis/nisplus-network.c: Likewise.
	* nss/nss_files/files-network.c: Likewise.
	* resolv/netdb.h: Define uint32_t if not already happened.
	Include bits/netdb.h.  Remove definition of struct netent.
	Change type of first parameter of getnetbyaddr_r and getnetbyaddr.
	* sysdeps/generic/stdint.h: Allow definition of uint32_t elsewhere.
	* sysdeps/generic/bits/netdb.h: New file.
	* sysdeps/unix/sysv/linux/alpha/bits/netdb.h: New file.

+++ include/netdb.h
@@ -31,7 +31,7 @@ extern int __getnetent_r (struct netent *__restrict __result_buf,
 			  struct netent **__restrict __result,
 			  int *__restrict __h_errnop);
 
-extern int __getnetbyaddr_r (unsigned long int __net, int __type,
+extern int __getnetbyaddr_r (uint32_t __net, int __type,
 			     struct netent *__restrict __result_buf,
 			     char *__restrict __buf, size_t __buflen,
 			     struct netent **__restrict __result,
